+++ src/py/reactpy/reactpy/backend/sanic.py
@@ -24,8 +24,8 @@ from reactpy.backend._common import (
     safe_web_modules_dir_path,
     serve_with_uvicorn,
 )
-from reactpy.backend.hooks import ConnectionContext
-from reactpy.backend.hooks import use_connection as _use_connection
+from reactpy.core.hooks import ConnectionContext
+from reactpy.core.hooks import use_connection as _use_connection
 from reactpy.backend.types import Connection, Location
 from reactpy.core.layout import Layout
 from reactpy.core.serve import RecvCoroutine, SendCoroutine, Stop, serve_layout
