+++ src/py/reactpy/reactpy/core/hooks.py
@@ -1,4 +1,6 @@
 from __future__ import annotations
+from collections.abc import MutableMapping
+from reactpy.backend.types import Connection, Location
 
 import asyncio
 from collections.abc import Coroutine, Sequence
@@ -227,6 +229,19 @@ def create_context(default_value: _Type) -> Context[_Type]:
     return context
 
 
+# backend implementations should establish this context at the root of an app
+ConnectionContext: Context[Connection[Any] | None] = create_context(None)
+
+
+def use_connection() -> Connection[Any]:
+    """Get the current :class:`~reactpy.backend.types.Connection`."""
+    conn = use_context(ConnectionContext)
+    if conn is None:  # nocov
+        msg = "No backend established a connection."
+        raise RuntimeError(msg)
+    return conn
+
+
 def use_context(context: Context[_Type]) -> _Type:
     """Get the current value for the given context type.
 
@@ -246,6 +261,16 @@ def use_context(context: Context[_Type]) -> _Type:
         return cast(_Type, context.__kwdefaults__["value"])
 
     return provider.value
+
+
+def use_scope() -> MutableMapping[str, Any]:
+    """Get the current :class:`~reactpy.backend.types.Connection`'s scope."""
+    return use_connection().scope
+
+
+def use_location() -> Location:
+    """Get the current :class:`~reactpy.backend.types.Connection`'s location."""
+    return use_connection().location
 
 
 class _ContextProvider(Generic[_Type]):
